+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod config;
 pub mod merkle_tree;
-// pub mod poseidon;
 pub mod recursion;
 // pub mod rollup;
 pub mod sparse_merkle_tree;
