+++ src/transaction/gadgets/block_header/mod.rs
@@ -5,7 +5,7 @@ use plonky2::{
     plonk::{circuit_builder::CircuitBuilder, config::AlgebraicHasher},
 };
 
-use crate::utils::gadgets::poseidon::poseidon_two_to_one;
+use crate::utils::gadgets::hash::poseidon_two_to_one;
 
 use super::super::block_header::BlockHeader;
